# static/openapi.yaml
openapi: 3.1.0
info:
  title: CRM REST API
  version: 1.0.0
  summary: Документация REST API CRM
  description: |
    Описание REST API для CRM-приложения. Большинство эндпоинтов требует
    авторизации через cookie-сессию и защиты CSRF для методов изменения данных.

    В проекте используется cookie-сессия Flask-Login и CSRF-защита Flask-WTF.
    Для POST/PUT/DELETE запросов добавляйте заголовок `X-CSRFToken` с актуальным токеном.
    Получить новый токен можно через `GET /refresh_csrf`.
servers:
  - url: /
    description: Текущий сервер
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Cookie-сессия Flask-Login (`session`)
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRFToken
      description: CSRF-токен Flask-WTF. Получайте через GET /refresh_csrf
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      description: Максимальное количество элементов (по умолчанию 50, не более 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
    Offset:
      name: offset
      in: query
      required: false
      description: Смещение выборки (по умолчанию 0)
      schema:
        type: integer
        minimum: 0
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: Проверка ETag для условного GET
      schema:
        type: string
    IfModifiedSince:
      name: If-Modified-Since
      in: header
      required: false
      description: Проверка даты изменения для условного GET
      schema:
        type: string
        format: date-time
  schemas:
...
    Object:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: [string, 'null']
        customer:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        created_at:
          type: [string, 'null']
          format: date-time
      required: [id, name]
...
    ContractorBrief:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required: [id, name]
    RequestStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, ACCEPTED, ASSIGNED, DONE, REJECTED, NEED_INFO]
    RequestSummary:
      type: object
      properties:
        id:
          type: integer
        status:
          $ref: '#/components/schemas/RequestStatus'
        created_at:
          type: string
          format: date-time
        manufacturers:
          type: array
          items:
            type: string
        contractors:
          type: array
          items:
            $ref: '#/components/schemas/ContractorBrief'
      required: [id, status, created_at, manufacturers, contractors]
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/get_all_objects:
    get:
      tags: [Objects]
      summary: Список объектов
      description: Получить все объекты (для автокомплита и справочников).
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
      responses:
        '200':
          description: Успех
          headers:
            ETag:
              description: Текущее значение ETag
              schema:
                type: string
            Last-Modified:
              description: Дата последнего изменения
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Object'
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: uri
                      next:
                        type: [string, 'null']
                        format: uri
                      related:
                        type: [string, 'null']
                        format: uri
                required: [data, links]
              examples:
                sample:
                  value:
                    data:
                      - id: 1
                        name: 'Объект 1'
                        address: 'ул. Пушкина, 1'
                        customer: 'ООО Ридан'
                    links:
                      self: 'https://example.com/api/v1/get_all_objects?limit=50&offset=0'
                      next: null
                      related: 'https://example.com/api/v1/get_all_contractors'
        '304':
          description: Не изменилось
        '500':
          description: Ошибка сервера
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: {}
                  links:
                    type: object
  /api/v1/get_all_contractors:
    get:
      tags: [Contractors]
      summary: Список подрядчиков
      description: Получить всех подрядчиков (для автокомплита и справочников).
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
      responses:
        '200':
          description: Успех
          headers:
            ETag:
              description: Текущее значение ETag
              schema:
                type: string
            Last-Modified:
              description: Дата последнего изменения
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContractorBrief'
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: uri
                      next:
                        type: [string, 'null']
                        format: uri
                      related:
                        type: [string, 'null']
                        format: uri
                required: [data, links]
              examples:
                sample:
                  value:
                    data:
                      - id: 10
                        name: 'ООО ТеплоМонтаж'
                    links:
                      self: 'https://example.com/api/v1/get_all_contractors?limit=50&offset=0'
                      next: null
                      related: 'https://example.com/api/v1/get_all_objects'
        '304':
          description: Не изменилось
        '500':
          description: Ошибка сервера
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: {}
                  links:
                    type: object
  '/api/v1/requests/by_contractor/{contractor_id}':
    get:
      tags: [Requests]
      summary: Заявки по подрядчику
      description: Возвращает список заявок, связанных с указанным подрядчиком.
      parameters:
        - name: contractor_id
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
      responses:
        '200':
          description: Список заявок
          headers:
            ETag:
              description: Текущее значение ETag
              schema:
                type: string
            Last-Modified:
              description: Дата последнего изменения
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RequestSummary'
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: uri
                      next:
                        type: [string, 'null']
                        format: uri
                      related:
                        type: [string, 'null']
                        format: uri
                required: [data, links]
              examples:
                sample:
                  value:
                    data:
                      - id: 101
                        status: OPEN
                        created_at: '2025-01-10T12:34:56Z'
                        manufacturers: ['Пульсар', 'Евра']
                        contractors:
                          - { id: 55, name: 'ООО ТеплоМонтаж' }
                    links:
                      self: 'https://example.com/api/v1/requests/by_contractor/1?limit=50&offset=0'
                      next: null
                      related: 'https://example.com/api/v1/get_all_contractors'
        '304':
          description: Не изменилось
        '500':
          description: Ошибка сервера
  '/api/v1/requests/by_object/{object_id}':
    get:
      tags: [Requests]
      summary: Заявки по объекту
      description: Возвращает список заявок для указанного объекта.
      parameters:
        - name: object_id
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
      responses:
        '200':
          description: Список заявок
          headers:
            ETag:
              description: Текущее значение ETag
              schema:
                type: string
            Last-Modified:
              description: Дата последнего изменения
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RequestSummary'
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: uri
                      next:
                        type: [string, 'null']
                        format: uri
                      related:
                        type: [string, 'null']
                        format: uri
                required: [data, links]
        '304':
          description: Не изменилось
        '500':
          description: Ошибка сервера
...